Rebased for split views, added error view to util
